+++ R/dataDiscretize.R
@@ -103,16 +103,16 @@ dataDiscretize <- function(data, classBoundaries=NULL, classStates=NULL, method=
             stop('Number of bins must match the number of states')
         }
         if(classBoundaries < 2 | abs(classBoundaries - round(classBoundaries)) > 0 ){
-            stop('"classBoundaries" must be an integer greater than 1, 
-                 or a vector of values to be used as class boundaries')
+            stop('"classBoundaries" must be an integer greater than 1, ',
+                 'or a vector of values to be used as class boundaries')
         }
         if(method == "quantile"){
             classBoundaries <- stats::quantile(data, probs=cumsum(rep(1/classBoundaries, classBoundaries-1)), 
                                                na.rm=TRUE, names = FALSE)
             classBoundaries <- c(mn, classBoundaries, mx)
             if(any(duplicated(classBoundaries))){
-                stop('Non unique quantile separators (a single value may cover a substantial fraction of the data).',
-                     ' Please specify a vector of class boundaries instead.')
+                stop('Non unique quantile separators (a single value may cover a substantial fraction of the data). ',
+                     'Please specify a vector of class boundaries instead.')
             }
         } 
         if(method == "equal"){